requirements.txt
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
PyQt5>=5.15.0
pillow>=8.0.0
numpy>=1.21.0
timm>=0.6.0
transformers>=4.20.0
psutil>=5.8.0
